Added some programs

+++ CountSubstrings.cpp
@@ -1,11 +1,13 @@
+// https://www.codechef.com/problems/CSUB
+
 #include <iostream>
 using namespace std;
 
-int countOnes() {
+unsigned long long int countOnes() {
     int num;
     char c;
     cin >> num;
-    int count = 0;
+    unsigned long long int count = 0;
     while(num--) {
         cin >> c;
         if(c == '1') {
@@ -19,7 +21,7 @@ int main() {
     int tc;
     cin >> tc;
     while(tc--) {
-        int count = countOnes();
+        unsigned long long int count = countOnes();
         count = (count * (count + 1)) / 2;
         cout << count << endl;
     }
